refactor: Retrofit 관련 코드 정리 및 분리, AuthProvider 패키지 변경
