+++ app/src/main/java/com/jukco/waitforme/data/mock/MockAuthProvider.kt
@@ -1,7 +1,7 @@
 package com.jukco.waitforme.data.mock
 
 import androidx.compose.runtime.Composable
-import com.jukco.waitforme.data.repository.AuthProvider
+import com.jukco.waitforme.data.auth.AuthProvider
 import com.jukco.waitforme.ui.sign.social.AuthUiProvider
 import com.jukco.waitforme.ui.sign.social.MockAuthUiProvider
