+++ app/src/main/java/com/jukco/waitforme/data/repository/AppContainer.kt
@@ -2,17 +2,17 @@ package com.jukco.waitforme.data.repository
 
 import android.content.Context
 import androidx.credentials.CredentialManager
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.jukco.waitforme.BuildConfig
+import com.jukco.waitforme.data.RetrofitUtil.signApi
+import com.jukco.waitforme.data.RetrofitUtil.storeApi
+import com.jukco.waitforme.data.RetrofitUtil.userApi
+import com.jukco.waitforme.data.auth.AuthProvider
+import com.jukco.waitforme.data.auth.GoogleAuthProvider
+import com.jukco.waitforme.data.auth.KakaoAuthProvider
+import com.jukco.waitforme.data.auth.NaverAuthProvider
 import com.jukco.waitforme.data.mock.MockSignRepository
-import com.jukco.waitforme.data.mock.MockUserRepository
 import com.jukco.waitforme.data.mock.MockStoreRepository
-import com.jukco.waitforme.data.network.api.SignApi
-import com.jukco.waitforme.data.network.api.StoreApi
-import com.jukco.waitforme.data.network.api.UserApi
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import retrofit2.Retrofit
+import com.jukco.waitforme.data.mock.MockUserRepository
 
 interface AppContainer {
     val tokenManager: TokenManager
@@ -25,6 +25,8 @@ interface AppContainer {
 }
 
 class DefaultContainer(context: Context) : AppContainer {
+    private val credentialManager = CredentialManager.create(context)
+
     override val tokenManager: TokenManager by lazy {
         TokenManagerImplementation(context.signInData)
     }
@@ -37,22 +39,6 @@ class DefaultContainer(context: Context) : AppContainer {
     }
     override val naverAuthProvider: AuthProvider by lazy {
         NaverAuthProvider()
-    }
-
-    private val retrofit = Retrofit.Builder()
-        .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-        .baseUrl(BuildConfig.SERVER_URL)
-        .build()
-    private val credentialManager = CredentialManager.create(context)
-
-    private val storeApi: StoreApi by lazy {
-        retrofit.create(StoreApi::class.java)
-    }
-    private val signApi: SignApi by lazy {
-        retrofit.create(SignApi::class.java)
-    }
-    private val userApi: UserApi by lazy {
-        retrofit.create(UserApi::class.java)
     }
 
     // TODO : 서버 연결 전까지 임시 Repository
