+++ app/src/main/java/com/jukco/waitforme/ui/components/SocialSignContainer.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
-import com.jukco.waitforme.data.repository.AuthProvider
+import com.jukco.waitforme.data.auth.AuthProvider
 import com.jukco.waitforme.ui.sign.sign_up.SignUpDto
 import kotlinx.coroutines.launch
